% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gd_es_municipio}
\alias{gd_es_municipio}
\title{\code{gd_es_municipio}}
\format{
A \code{geodimension}.
}
\source{
\url{https://www.ign.es}, \url{https://www.mapa.gob.es}
}
\usage{
gd_es_municipio
}
\description{
\code{geodimension} obtained from vector layers over Spain.
}
\details{
It includes the levels municipality, livestock region, agricultural region,
province, autonomous community and country.
}
\keyword{datasets}
