% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gd_us_place}
\alias{gd_us_place}
\title{\code{gd_us_place}}
\format{
A \code{geodimension}.
}
\source{
\url{https://www.census.gov}
}
\usage{
gd_us_place
}
\description{
\code{geodimension} obtained from vector layers over USA.
}
\details{
It includes the levels place, county, Congressional District (116th
Congress), State Legislative Districts - Upper Chamber, State Legislative
Districts - Lower Chamber, state, region, division and nation.
}
\keyword{datasets}
